race packaging and adds elf
